Adding example of stamped lock and free lock

+++ Spring/SpringBootAdvanced/src/main/java/ar/com/javacuriosities/springboot/model/Role.java
@@ -17,10 +17,10 @@ public class Role {
     @Column(name="role")
     private String role;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
     public String getRole() {
